numpy depreciations of np.float np.bool

+++ merlin/analysis/preprocess.py
@@ -118,7 +118,7 @@ class CAREPreprocess(Preprocess):
         hpImage = imagefilters.high_pass_filter(inputImage,
                                                 highPassFilterSize,
                                                 self._highPassSigma)
-        return hpImage.astype(np.float)
+        return hpImage.astype('float16')
     
     def _run_analysis(self, fragmentIndex):
     
@@ -232,7 +232,7 @@ class DeconvolutionPreprocess(Preprocess):
         hpImage = imagefilters.high_pass_filter(inputImage,
                                                 highPassFilterSize,
                                                 self._highPassSigma)
-        return hpImage.astype(np.float)
+        return hpImage.astype('float16')
 
     def _run_analysis(self, fragmentIndex):
